Added ability to change the size of the spectral deminsion to be fit to, via the parameter optcut which specifies the number of pixels on either side of the center pf the array. This often produces erroneous results and is not currently recommendd to be changed. The default is 40, inline with an 81 pixel-long spectral axis.
Created a 1D PRF function to be called in all the areas it was previously defined (decluttering code).

Changed std of prf fit model in the spectral axis to be 3 pixels instead of 1.

Changed name of noise_array to cutout_forfit (and subsequently cut_cutout_forfit for optimzed cutting.) This is the array of data to be fit to, and now has a more logical name.

Least squares method has  been found to have many bugs. The parameter 'prf_fitmethod' can be used to change between curve_fit and least_sqaures, and defaults to the former. It is not recommended to use least_squares.

If PRF fitting does not work, make sure optcut and prf_fitmethod are specified.